+++ android/src/main/java/com/lib/flutter_blue_plus/FlutterBluePlusPlugin.java
@@ -223,6 +223,7 @@ public class FlutterBluePlusPlugin implements
         if (mBluetoothAdapter != null && mIsScanning) {
             BluetoothLeScanner scanner = mBluetoothAdapter.getBluetoothLeScanner();
             if (scanner != null) {
+                log(LogLevel.DEBUG, "calling stopScan (OnDetachedFromEngine)");
                 scanner.stopScan(getScanCallback());
                 mIsScanning = false;
             }
@@ -340,6 +341,7 @@ public class FlutterBluePlusPlugin implements
                     // stop scanning
                     BluetoothLeScanner scanner = mBluetoothAdapter.getBluetoothLeScanner();
                     if(scanner != null && mIsScanning) {
+                        log(LogLevel.DEBUG, "calling stopScan (flutterRestart)");
                         scanner.stopScan(getScanCallback());
                         mIsScanning = false;
                     }
@@ -621,7 +623,7 @@ public class FlutterBluePlusPlugin implements
                         // service data
                         for (int i = 0; i < withServiceData.size(); i++) {
                             HashMap<String, Object> m = (HashMap<String, Object>) withServiceData.get(i);
-                            ParcelUuid s = ParcelUuid.fromString((String) m.get("service"));
+                            ParcelUuid s = ParcelUuid.fromString(uuid128((String) m.get("service")));
                             byte[] mdata =             hexToBytes((String) m.get("data"));
                             byte[] mask =              hexToBytes((String) m.get("mask"));
                             ScanFilter f = null;
@@ -855,10 +857,10 @@ public class FlutterBluePlusPlugin implements
                 {
                     // see: BmReadCharacteristicRequest
                     HashMap<String, Object> data = call.arguments();
-                    String remoteId =             (String) data.get("remote_id");
-                    String serviceUuid =          (String) data.get("service_uuid");
-                    String secondaryServiceUuid = (String) data.get("secondary_service_uuid");
-                    String characteristicUuid =   (String) data.get("characteristic_uuid");
+                    String remoteId =           (String) data.get("remote_id");
+                    String serviceUuid =        (String) data.get("service_uuid");
+                    String characteristicUuid = (String) data.get("characteristic_uuid");
+                    String primaryServiceUuid = (String) data.get("primary_service_uuid");
 
                     // check connection
                     BluetoothGatt gatt = mConnectedDevices.get(remoteId);
@@ -871,7 +873,7 @@ public class FlutterBluePlusPlugin implements
                     waitIfBonding();
 
                     // find characteristic
-                    ChrFound found = locateCharacteristic(gatt, serviceUuid, secondaryServiceUuid, characteristicUuid);
+                    ChrFound found = locateCharacteristic(gatt, serviceUuid, characteristicUuid, primaryServiceUuid);
                     if (found.error != null) {
                         result.error("readCharacteristic", found.error, null);
                         break;
@@ -901,13 +903,13 @@ public class FlutterBluePlusPlugin implements
                 {
                     // see: BmWriteCharacteristicRequest
                     HashMap<String, Object> data = call.arguments();
-                    String remoteId =             (String) data.get("remote_id");
-                    String serviceUuid =          (String) data.get("service_uuid");
-                    String secondaryServiceUuid = (String) data.get("secondary_service_uuid");
-                    String characteristicUuid =   (String) data.get("characteristic_uuid");
-                    String value =                (String) data.get("value");
-                    int writeTypeInt =               (int) data.get("write_type");
-                    boolean allowLongWrite =        ((int) data.get("allow_long_write")) != 0;
+                    String remoteId =           (String) data.get("remote_id");
+                    String serviceUuid =        (String) data.get("service_uuid");
+                    String characteristicUuid = (String) data.get("characteristic_uuid");
+                    String primaryServiceUuid = (String) data.get("primary_service_uuid");
+                    String value =              (String) data.get("value");
+                    int writeTypeInt =             (int) data.get("write_type");
+                    boolean allowLongWrite =      ((int) data.get("allow_long_write")) != 0;
 
                     int writeType = writeTypeInt == 0 ?
                         BluetoothGattCharacteristic.WRITE_TYPE_DEFAULT :
@@ -924,7 +926,7 @@ public class FlutterBluePlusPlugin implements
                     waitIfBonding();
 
                     // find characteristic
-                    ChrFound found = locateCharacteristic(gatt, serviceUuid, secondaryServiceUuid, characteristicUuid);
+                    ChrFound found = locateCharacteristic(gatt, serviceUuid, characteristicUuid, primaryServiceUuid);
                     if (found.error != null) {
                         result.error("writeCharacteristic", found.error, null);
                         break;
@@ -959,7 +961,8 @@ public class FlutterBluePlusPlugin implements
                     }
 
                     // remember the data we are writing
-                    String key = remoteId + ":" + serviceUuid + ":" + characteristicUuid;
+                    if (primaryServiceUuid == null) {primaryServiceUuid = "";}
+                    String key = remoteId + ":" + serviceUuid + ":" + characteristicUuid + ":" + primaryServiceUuid;
                     mWriteChr.put(key, value);
 
                     // write characteristic
@@ -998,11 +1001,11 @@ public class FlutterBluePlusPlugin implements
                 {
                     // see: BmReadDescriptorRequest
                     HashMap<String, Object> data = call.arguments();
-                    String remoteId =             (String) data.get("remote_id");
-                    String serviceUuid =          (String) data.get("service_uuid");
-                    String secondaryServiceUuid = (String) data.get("secondary_service_uuid");
-                    String characteristicUuid =   (String) data.get("characteristic_uuid");
-                    String descriptorUuid =       (String) data.get("descriptor_uuid");
+                    String remoteId =           (String) data.get("remote_id");
+                    String serviceUuid =        (String) data.get("service_uuid");
+                    String characteristicUuid = (String) data.get("characteristic_uuid");
+                    String descriptorUuid =     (String) data.get("descriptor_uuid");
+                    String primaryServiceUuid = (String) data.get("primary_service_uuid");
 
                     // check connection
                     BluetoothGatt gatt = mConnectedDevices.get(remoteId);
@@ -1015,7 +1018,7 @@ public class FlutterBluePlusPlugin implements
                     waitIfBonding();
 
                     // find characteristic
-                    ChrFound found = locateCharacteristic(gatt, serviceUuid, secondaryServiceUuid, characteristicUuid);
+                    ChrFound found = locateCharacteristic(gatt, serviceUuid, characteristicUuid, primaryServiceUuid);
                     if (found.error != null) {
                         result.error("readDescriptor", found.error, null);
                         break;
@@ -1045,12 +1048,12 @@ public class FlutterBluePlusPlugin implements
                 {
                     // see: BmWriteDescriptorRequest
                     HashMap<String, Object> data = call.arguments();
-                    String remoteId =             (String) data.get("remote_id");
-                    String serviceUuid =          (String) data.get("service_uuid");
-                    String secondaryServiceUuid = (String) data.get("secondary_service_uuid");
-                    String characteristicUuid =   (String) data.get("characteristic_uuid");
-                    String descriptorUuid =       (String) data.get("descriptor_uuid");
-                    String value =                (String) data.get("value");
+                    String remoteId =           (String) data.get("remote_id");
+                    String serviceUuid =        (String) data.get("service_uuid");
+                    String characteristicUuid = (String) data.get("characteristic_uuid");
+                    String descriptorUuid =     (String) data.get("descriptor_uuid");
+                    String primaryServiceUuid = (String) data.get("primary_service_uuid");
+                    String value =              (String) data.get("value");
 
                     // check connection
                     BluetoothGatt gatt = mConnectedDevices.get(remoteId);
@@ -1063,7 +1066,7 @@ public class FlutterBluePlusPlugin implements
                     waitIfBonding();
 
                     // find characteristic
-                    ChrFound found = locateCharacteristic(gatt, serviceUuid, secondaryServiceUuid, characteristicUuid);
+                    ChrFound found = locateCharacteristic(gatt, serviceUuid, characteristicUuid, primaryServiceUuid);
                     if (found.error != null) {
                         result.error("writeDescriptor", found.error, null);
                         break;
@@ -1089,7 +1092,8 @@ public class FlutterBluePlusPlugin implements
                     }
 
                     // remember the data we are writing
-                    String key = remoteId + ":" + serviceUuid + ":" + characteristicUuid + ":" + descriptorUuid;
+                    if (primaryServiceUuid == null) {primaryServiceUuid = "";}
+                    String key = remoteId + ":" + serviceUuid + ":" + characteristicUuid + ":" + descriptorUuid + ":" + primaryServiceUuid;
                     mWriteDesc.put(key, value);
 
                     // write descriptor
@@ -1125,12 +1129,12 @@ public class FlutterBluePlusPlugin implements
                 {
                     // see: BmSetNotifyValueRequest
                     HashMap<String, Object> data = call.arguments();
-                    String remoteId =             (String) data.get("remote_id");
-                    String serviceUuid =          (String) data.get("service_uuid");
-                    String secondaryServiceUuid = (String) data.get("secondary_service_uuid");
-                    String characteristicUuid =   (String) data.get("characteristic_uuid");
-                    boolean forceIndications =   (boolean) data.get("force_indications");
-                    boolean enable =             (boolean) data.get("enable");
+                    String remoteId =            (String) data.get("remote_id");
+                    String serviceUuid =         (String) data.get("service_uuid");
+                    String characteristicUuid =  (String) data.get("characteristic_uuid");
+                    String primaryServiceUuid = (String) data.get("primary_service_uuid");
+                    boolean forceIndications =  (boolean) data.get("force_indications");
+                    boolean enable =            (boolean) data.get("enable");
 
                     // check connection
                     BluetoothGatt gatt = mConnectedDevices.get(remoteId);
@@ -1143,7 +1147,7 @@ public class FlutterBluePlusPlugin implements
                     waitIfBonding();
 
                     // find characteristic
-                    ChrFound found = locateCharacteristic(gatt, serviceUuid, secondaryServiceUuid, characteristicUuid);
+                    ChrFound found = locateCharacteristic(gatt, serviceUuid, characteristicUuid, primaryServiceUuid);
                     if (found.error != null) {
                         result.error("setNotifyValue", found.error, null);
                         break;
@@ -1201,7 +1205,8 @@ public class FlutterBluePlusPlugin implements
                     }
 
                     // remember the data we are writing
-                    String key = remoteId + ":" + serviceUuid + ":" + characteristicUuid + ":" + CCCD;
+                    if (primaryServiceUuid == null) {primaryServiceUuid = "";}
+                    String key = remoteId + ":" + serviceUuid + ":" + characteristicUuid + ":" + CCCD + ":" + primaryServiceUuid;
                     mWriteDesc.put(key, bytesToHex(descriptorValue));
 
                     // write descriptor
@@ -1808,22 +1813,30 @@ public class FlutterBluePlusPlugin implements
     }
 
     private ChrFound locateCharacteristic(BluetoothGatt gatt,
-                                                 String serviceId,
-                                                 String secondaryServiceId,
-                                                 String characteristicId)
+                                                 String serviceUuid,
+                                                 String characteristicUuid, 
+                                                 String primaryServiceUuid)
     {
+        // remember this
+        boolean isSecondaryService = primaryServiceUuid != null;
+
+        // primary service
+        if (primaryServiceUuid == null) {
+            primaryServiceUuid = serviceUuid;
+        }
+
         // primary
-        BluetoothGattService primaryService = getServiceFromArray(serviceId, gatt.getServices());
+        BluetoothGattService primaryService = getServiceFromArray(primaryServiceUuid, gatt.getServices());
         if(primaryService == null) {
-            return new ChrFound(null, "service not found '" + serviceId + "'");
+            return new ChrFound(null, "primary service not found '" + primaryServiceUuid + "'");
         }
 
         // secondary
         BluetoothGattService secondaryService = null;
-        if(secondaryServiceId != null && secondaryServiceId.length() > 0) {
-            secondaryService = getServiceFromArray(serviceId, primaryService.getIncludedServices());
+        if(isSecondaryService) {
+            secondaryService = getServiceFromArray(serviceUuid, primaryService.getIncludedServices());
             if(secondaryService == null) {
-                return new ChrFound(null, "secondaryService not found '" + secondaryServiceId + "'");
+                return new ChrFound(null, "secondary service not found '" + serviceUuid + "'");
             }
         }
 
@@ -1831,10 +1844,10 @@ public class FlutterBluePlusPlugin implements
         BluetoothGattService service = (secondaryService != null) ? secondaryService : primaryService;
 
         // characteristic
-        BluetoothGattCharacteristic characteristic = getCharacteristicFromArray(characteristicId, service.getCharacteristics());
+        BluetoothGattCharacteristic characteristic = getCharacteristicFromArray(characteristicUuid, service.getCharacteristics());
         if(characteristic == null) {
             return new ChrFound(null, "characteristic not found in service " +
-                "(chr: '" + characteristicId + "' svc: '" + serviceId + "')");
+                "(chr: '" + characteristicUuid + "' svc: '" + serviceUuid + "')");
         }
 
         return new ChrFound(characteristic, null);
@@ -2100,13 +2113,14 @@ public class FlutterBluePlusPlugin implements
 
             log(LogLevel.DEBUG, "OnAdapterStateChanged: " + adapterStateString(adapterState));
 
-            // stop scanning when adapter is turned off.
+            // stop scanning when adapter is turned back on.
             // Otherwise, scanning automatically resumes when the adapter is
             // turned back on. I don't think most users expect that.
-            if (adapterState != BluetoothAdapter.STATE_ON) {
+            if (adapterState == BluetoothAdapter.STATE_ON) {
                 if (mBluetoothAdapter != null && mIsScanning) {
                     BluetoothLeScanner scanner = mBluetoothAdapter.getBluetoothLeScanner();
                     if (scanner != null) {
+                        log(LogLevel.DEBUG, "calling stopScan (Bluetooth Restarted)");
                         scanner.stopScan(getScanCallback());
                         mIsScanning = false;
                     }
@@ -2431,7 +2445,10 @@ public class FlutterBluePlusPlugin implements
 
             List<Object> services = new ArrayList<Object>();
             for(BluetoothGattService s : gatt.getServices()) {
-                services.add(bmBluetoothService(gatt.getDevice(), s, gatt));
+                services.add(bmBluetoothService(gatt.getDevice(), s, null, gatt));
+                for(BluetoothGattService s2 : s.getIncludedServices()) {
+                    services.add(bmBluetoothService(gatt.getDevice(), s2, s, gatt));
+                }
             }
 
             // see: BmDiscoverServicesResult
@@ -2448,10 +2465,8 @@ public class FlutterBluePlusPlugin implements
         // called for both notifications & reads
         public void onCharacteristicReceived(BluetoothGatt gatt, BluetoothGattCharacteristic characteristic, byte[] value, int status)
         {
-            ServicePair pair = getServicePair(gatt, characteristic);
-
             // GATT Service?
-            if (uuidStr(pair.primary) == "1800") {
+            if (uuidStr(characteristic.getService().getUuid()) == "1800") {
 
                 // services changed
                 if (uuidStr(characteristic.getUuid()) == "2A05") {
@@ -2460,18 +2475,21 @@ public class FlutterBluePlusPlugin implements
                 }
             }
 
+            // has associated primary service?
+            BluetoothGattService primaryService = getPrimaryService(gatt, characteristic);
+
             // see: BmCharacteristicData
             HashMap<String, Object> response = new HashMap<>();
             response.put("remote_id", gatt.getDevice().getAddress());
-            response.put("service_uuid", uuidStr(pair.primary));
-            if (pair.secondary != null) {
-                response.put("secondary_service_uuid", uuidStr(pair.secondary));
-            }
+            response.put("service_uuid", uuidStr(characteristic.getService().getUuid()));
             response.put("characteristic_uuid", uuidStr(characteristic.getUuid()));
             response.put("value", bytesToHex(value));
             response.put("success", status == BluetoothGatt.GATT_SUCCESS ? 1 : 0);
             response.put("error_code", status);
             response.put("error_string", gattErrorString(status));
+            if (primaryService != null) {
+                response.put("primary_service_uuid", uuidStr(primaryService.getUuid()));
+            }
 
             invokeMethodUIThread("OnCharacteristicReceived", response);
         }
@@ -2512,16 +2530,17 @@ public class FlutterBluePlusPlugin implements
             // in android's internal buffer. When the buffer is full, it delays calling onCharacteristicWrite
             // until there is at least ~50% free space again.
 
-            ServicePair pair = getServicePair(gatt, characteristic);
+            // has associated primary service?
+            BluetoothGattService primaryService = getPrimaryService(gatt, characteristic);
 
             // for convenience
             String remoteId = gatt.getDevice().getAddress();
-            String serviceUuid = uuidStr(pair.primary);
-            String secondaryServiceUuid = pair.secondary != null ? uuidStr(pair.secondary) : null;
+            String serviceUuid = uuidStr(characteristic.getService().getUuid());
             String characteristicUuid = uuidStr(characteristic.getUuid());
+            String primaryServiceUuid = primaryService != null ? uuidStr(primaryService.getUuid()) : "";
 
             // what data did we write?
-            String key = remoteId + ":" + serviceUuid + ":" + characteristicUuid;
+            String key = remoteId + ":" + serviceUuid + ":" + characteristicUuid + ":" + primaryServiceUuid;
             String value = mWriteChr.get(key) != null ? mWriteChr.get(key) : "";
             mWriteChr.remove(key);
 
@@ -2529,15 +2548,15 @@ public class FlutterBluePlusPlugin implements
             HashMap<String, Object> response = new HashMap<>();
             response.put("remote_id", remoteId);
             response.put("service_uuid", serviceUuid);
-            if (secondaryServiceUuid != null) {
-                response.put("secondary_service_uuid", secondaryServiceUuid);
-            }
             response.put("characteristic_uuid", characteristicUuid);
             response.put("value", value);
             response.put("success", status == BluetoothGatt.GATT_SUCCESS ? 1 : 0);
             response.put("error_code", status);
             response.put("error_string", gattErrorString(status));
-
+            if (primaryService != null) {
+                response.put("primary_service_uuid", uuidStr(primaryService.getUuid()));
+            }
+            
             invokeMethodUIThread("OnCharacteristicWritten", response);
         }
 
@@ -2551,21 +2570,22 @@ public class FlutterBluePlusPlugin implements
             log(level, "  desc: " + uuidStr(descriptor.getUuid()));
             log(level, "  status: " + gattErrorString(status) + " (" + status + ")");
 
-            ServicePair pair = getServicePair(gatt, descriptor.getCharacteristic());
+            // has associated primary service?
+            BluetoothGattService primaryService = getPrimaryService(gatt, descriptor.getCharacteristic());
 
             // see: BmDescriptorData
             HashMap<String, Object> response = new HashMap<>();
             response.put("remote_id", gatt.getDevice().getAddress());
-            response.put("service_uuid", uuidStr(pair.primary));
-            if (pair.secondary != null) {
-                response.put("secondary_service_uuid", uuidStr(pair.secondary));
-            }
+            response.put("service_uuid", uuidStr(descriptor.getCharacteristic().getService().getUuid()));
             response.put("characteristic_uuid", uuidStr(descriptor.getCharacteristic().getUuid()));
             response.put("descriptor_uuid", uuidStr(descriptor.getUuid()));
             response.put("value", bytesToHex(value));
             response.put("success", status == BluetoothGatt.GATT_SUCCESS ? 1 : 0);
             response.put("error_code", status);
             response.put("error_string", gattErrorString(status));
+            if (primaryService != null) {
+                response.put("primary_service_uuid", uuidStr(primaryService.getUuid()));
+            }
 
             invokeMethodUIThread("OnDescriptorRead", response);
         }
@@ -2579,17 +2599,18 @@ public class FlutterBluePlusPlugin implements
             log(level, "  desc: " + uuidStr(descriptor.getUuid()));
             log(level, "  status: " + gattErrorString(status) + " (" + status + ")");
 
-            ServicePair pair = getServicePair(gatt, descriptor.getCharacteristic());
+            // has associated primary service?
+            BluetoothGattService primaryService = getPrimaryService(gatt, descriptor.getCharacteristic());
 
             // for convenience
             String remoteId = gatt.getDevice().getAddress();
-            String serviceUuid = uuidStr(pair.primary);
-            String secondaryServiceUuid = pair.secondary != null ? uuidStr(pair.secondary) : null;
+            String serviceUuid = uuidStr(descriptor.getCharacteristic().getService().getUuid());
             String characteristicUuid = uuidStr(descriptor.getCharacteristic().getUuid());
             String descriptorUuid = uuidStr(descriptor.getUuid());
+            String primaryServiceUuid = primaryService != null ? uuidStr(primaryService.getUuid()) : "";
 
             // what data did we write?
-            String key = remoteId + ":" + serviceUuid + ":" + characteristicUuid + ":" + descriptorUuid;
+            String key = remoteId + ":" + serviceUuid + ":" + characteristicUuid + ":" + descriptorUuid + ":" + primaryServiceUuid;
             String value = mWriteDesc.get(key) != null ? mWriteDesc.get(key) : "";
             mWriteDesc.remove(key);
 
@@ -2597,15 +2618,15 @@ public class FlutterBluePlusPlugin implements
             HashMap<String, Object> response = new HashMap<>();
             response.put("remote_id", remoteId);
             response.put("service_uuid", serviceUuid);
-            if (secondaryServiceUuid != null) {
-                response.put("secondary_service_uuid", secondaryServiceUuid);
-            }
             response.put("characteristic_uuid", characteristicUuid);
             response.put("descriptor_uuid", descriptorUuid);
             response.put("value", value);
             response.put("success", status == BluetoothGatt.GATT_SUCCESS ? 1 : 0);
             response.put("error_code", status);
             response.put("error_string", gattErrorString(status));
+            if (primaryService != null) {
+                response.put("primary_service_uuid", uuidStr(primaryService.getUuid()));
+            }
 
             invokeMethodUIThread("OnDescriptorWritten", response);
         }
@@ -2779,61 +2800,66 @@ public class FlutterBluePlusPlugin implements
         return map;
     }
 
-    HashMap<String, Object> bmBluetoothService(BluetoothDevice device, BluetoothGattService service, BluetoothGatt gatt) {
+    HashMap<String, Object> bmBluetoothService(
+        BluetoothDevice device, 
+        BluetoothGattService service,
+        BluetoothGattService primaryService,
+        BluetoothGatt gatt)
+    {
 
         List<Object> characteristics = new ArrayList<Object>();
         for(BluetoothGattCharacteristic c : service.getCharacteristics()) {
             characteristics.add(bmBluetoothCharacteristic(device, c, gatt));
         }
 
-        List<Object> includedServices = new ArrayList<Object>();
-        for(BluetoothGattService included : service.getIncludedServices()) {
-            // service includes itself?
-            if (included.getUuid().equals(service.getUuid())) {
-                continue; // skip, infinite recursion
-            }
-            includedServices.add(bmBluetoothService(device, included, gatt));
-        }
-
         // See: BmBluetoothService
         HashMap<String, Object> map = new HashMap<>();
         map.put("remote_id", device.getAddress());
         map.put("service_uuid", uuidStr(service.getUuid()));
-        map.put("is_primary", service.getType() == BluetoothGattService.SERVICE_TYPE_PRIMARY ? 1 : 0);
         map.put("characteristics", characteristics);
-        map.put("included_services", includedServices);
+        if (primaryService != null) {
+            map.put("primary_service_uuid", uuidStr(primaryService.getUuid()));
+        }
         return map;
     }
 
     HashMap<String, Object> bmBluetoothCharacteristic(BluetoothDevice device, BluetoothGattCharacteristic characteristic, BluetoothGatt gatt) {
 
-        ServicePair pair = getServicePair(gatt, characteristic);
-
         List<Object> descriptors = new ArrayList<Object>();
         for(BluetoothGattDescriptor d : characteristic.getDescriptors()) {
-            descriptors.add(bmBluetoothDescriptor(device, d));
+            descriptors.add(bmBluetoothDescriptor(device, d, gatt));
         }
+
+        // has associated primary service?
+        BluetoothGattService primaryService = getPrimaryService(gatt, characteristic);
 
         // See: BmBluetoothCharacteristic
         HashMap<String, Object> map = new HashMap<>();
         map.put("remote_id", device.getAddress());
-        map.put("service_uuid", uuidStr(pair.primary));
-        if (pair.secondary != null) {
-            map.put("secondary_service_uuid", uuidStr(pair.secondary));
-        }
+        map.put("service_uuid", uuidStr(characteristic.getService().getUuid()));
         map.put("characteristic_uuid", uuidStr(characteristic.getUuid()));
         map.put("descriptors", descriptors);
         map.put("properties", bmCharacteristicProperties(characteristic.getProperties()));
+        if (primaryService != null) {
+            map.put("primary_service_uuid", uuidStr(primaryService.getUuid()));
+        }
         return map;
     }
 
     // See: BmBluetoothDescriptor
-    HashMap<String, Object> bmBluetoothDescriptor(BluetoothDevice device, BluetoothGattDescriptor descriptor) {
+    HashMap<String, Object> bmBluetoothDescriptor(BluetoothDevice device, BluetoothGattDescriptor descriptor, BluetoothGatt gatt) {
+
+        // has associated primary service?
+        BluetoothGattService primaryService = getPrimaryService(gatt, descriptor.getCharacteristic());
+
         HashMap<String, Object> map = new HashMap<>();
         map.put("remote_id", device.getAddress());
         map.put("descriptor_uuid", uuidStr(descriptor.getUuid()));
         map.put("characteristic_uuid", uuidStr(descriptor.getCharacteristic().getUuid()));
         map.put("service_uuid", uuidStr(descriptor.getCharacteristic().getService().getUuid()));
+        if (primaryService != null) {
+            map.put("primary_service_uuid", uuidStr(primaryService.getUuid()));
+        }
         return map;
     }
 
@@ -2893,35 +2919,25 @@ public class FlutterBluePlusPlugin implements
         }
     }
 
-    public static class ServicePair {
-        public UUID primary;
-        public UUID secondary;
-    }
-
-    static ServicePair getServicePair(BluetoothGatt gatt, BluetoothGattCharacteristic characteristic) {
-
-        ServicePair result = new ServicePair();
+    static BluetoothGattService getPrimaryService(BluetoothGatt gatt, BluetoothGattCharacteristic characteristic) {
 
         BluetoothGattService service = characteristic.getService();
 
-        // is this a primary service?
+        // is this *already* a primary service? 
         if(service.getType() == BluetoothGattService.SERVICE_TYPE_PRIMARY) {
-            result.primary = service.getUuid();
-            return result;
+            return null;
         }
 
-        // Otherwise, iterate all services until we find the primary service
+        // Otherwise, iterate included services until we find the primary service
         for(BluetoothGattService primary : gatt.getServices()) {
             for(BluetoothGattService secondary : primary.getIncludedServices()) {
                 if(secondary.getUuid().equals(service.getUuid())) {
-                    result.primary = primary.getUuid();
-                    result.secondary = secondary.getUuid();
-                    return result;
+                    return primary;
                 }
             }
         }
 
-        return result;
+        return null;
     }
 
     //////////////////////////////////////////