+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 1.34.3
+* **[Fix]** `startScan` was not propogating errors
+
+## 1.34.2
+* **[Fix]** characteristic not found (regression 1.34.0)
+
+## 1.34.1
+* **[Fix]** Android: `SCAN_FAILED_ALREADY_STARTED` could sometimes occur after bluetooth restart
+
+## 1.34.0
+* **[Feature]** support `includedServices`, aka primary/secondary services
+* **[Fix]** Android: `withServiceData` scan filter was not working
+
+## 1.33.6
+* **[Feature]** verbose logs: make function results easier to see
+
 ## 1.33.5
 * **[Fix]** iOS: crash caused by 1-byte advertisements (#1022)